requirements.txt
alabaster==0.7.12
appdirs==1.4.3
astroid==2.3.3
astropy==4.3.1
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.9.1
black==19.10b0
bleach==3.3.0
certifi==2020.6.20
cffi==1.14.5
chardet==4.0.0
Click==7.0
cloudpickle==1.6.0
colorama==0.4.4
coverage==5.3
cryptography==3.4.7
cycler==0.10.0
cytoolz==0.11.0
dask==2.30.0
decorator==4.4.2
dnspython==1.16.0
docopt==0.6.2
docutils==0.17.1
fett==0.3.2
future==0.18.2
idna==2.10
imageio==2.6.1
imagesize==1.2.0
importlib-metadata==4.6.0
isort==4.3.21
jeepney==0.6.0
Jinja2==3.0.1
keyring==23.0.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
livereload==2.6.3
mapbox-vector-tile==1.2.1
MarkupSafe==2.0.1
matplotlib==3.1.2
mccabe==0.6.1
more-itertools==7.2.0
mypy-extensions==0.4.3
networkx==2.5
numpy==1.17.3
olefile==0.46
packaging==19.2
pathspec==0.6.0
Pillow==8.1.1
pkginfo==1.7.0
pluggy==0.13.0
protobuf==3.17.3
psutil==5.6.6
py==1.8.0
pyclipper==1.3.0
pycparser==2.20
pyerfa==2.0.0
Pygments==2.9.0
pylint==2.4.4
pymongo==3.11.4
pyparsing==2.4.5
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtWebEngine==5.12.1
pytest==5.2.3
pytest-arraydiff==0.3
pytest-astropy==0.6.0
pytest-astropy-header==0.1.1
pytest-cov==2.10.1
pytest-doctestplus==0.4.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.1
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.3.1
readme-renderer==29.0
regex==2019.11.1
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rstcheck==3.3.1
scikit-image==0.17.2
scipy==1.5.2
SecretStorage==3.3.1
Shapely==1.7.1
sharedmem==0.3.8
six==1.13.0
snooty-lextudio==1.9.2.dev0
snowballstemmer==2.1.0
Sphinx==4.0.2
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tifffile==2020.10.1
toml==0.10.0
toolz==0.11.1
tornado==6.0.3
tqdm==4.38.0
twine==3.4.1
typed-ast==1.4.0
typing-extensions==3.7.4.1
ujson==1.35
urllib3==1.26.6
watchdog==1.0.2
wcwidth==0.1.7
webencodings==0.5.1
wrapt==1.11.2
zipp==0.6.0
